Tried fixing scale for text
Getting font size to make sense if somewhat problematic, because
scaling is not straight forward. Font sizes that would be normal in
a text editor is way too big using TextMeshPro. This might be because
of the scaling of the canvas, but I suspect there is just no transferability
between TextMeshPro and the sizes of more conventional text processors

Closing the branch as this goes beyond what input actions are concerned with.