Little progress towards text curvature
Been digging through old TextMeshPro forums posts where people have
asked about text curvature. None of the examples found has worked out
of the box, but I am currently attempting to make the script provided
by the developer a few years back to work with the current version
of both unity and tmp.

An alternative to this could be to use paid solutions like CurvedUI($25)
or MegaFier($120), but I would really like to solve this myself if possible.
Going to continue trying for a little while, then contact professors
for guidance as to what to do if I can't make it work.
Committing the script I am trying to make work again, but little else
besides installing ProBuilder to create a curved surface within Unity
(avoids using Blender constantly, and possibility of changing the mesh
during runtime).

End-of-day commit.