End of day commit
Made some changes to work without a HMD present. Simply tick the XR support box again in Player settings and replace the Directional camera with the Camera rig again.

Figured out that the reason why my JSON always validated to null was because that I was using inconsistent names from the JSON compared to the objects that I created. It went over my head that serialization means consistency is key to reading and writing to the same file, thus making my files unreadable for Jsonutilities