Add write and read functionality of json to and from file.
This update makes it possible to write to and read from JSON files. Initially
this was recommended to be done through the use of UnityEngine.Jsonutilities, but
it seems like there is no support for lists/arrays yet. Through the power
of stackoverflow a helper class was found that manages this for me.
Writing to file will most likely not be used, but is kept in case
it becomes relevant when test data will be written to file.
As of now all the json handling is kept within the DataController.