Print out of some basic character info
It might not look like much, but this is a couple of hours of trying to
understand how to get basic info about the text that is being rendered.
Since I need to know the size of the letters, proabably at an individual
level, in terms of pixels or whatever translatable unit there is.