Add basic scaling of the screen based on distance away from center.
Using the "comfortable" viewing angles of people's eyes (approximated to be
about 30 degrees left and right, 20 degrees up and 12 degrees down) to scale
the screen based on how far away it is from the world center.
Some basic trigonometry is applied to calculate how big the screen needs to be
based on these viewing angles at the given distance to the screen.
Currently calculated in update, but can probably only calculated whenever
a the screen moves.