Scaling vision does not take reading task into consideration and I got sidetracked.
After experiencing how the dimensions fit into my vision in virtual reality,
I come to the realisation that what I've been looking at for reference points
were the comfortable range of human vision. This does not take into consideration
that reading should be done at the edges of human vision. Although it might be
comfortable to move your eyes 30 degrees to the left, it does not mean that
you can comfortably read at such an angle as well. Further literature needs to
be reviewed to find a good size of the screen.

The sidetrack comes from being able to scale the text. I should not have started
to tinker with this in this branch. I have however made it a bit easier to scale
the text while in virtual reality, where text size increases based on touch and
not touchpad presses. The closer to the edges the person in VR touches the touchpad,
the faster the text scales up and down.