Example code to curve text is now running.
I have now gotten the example script by the TMP developer to work. The forum
linked at the top of the script shows all the comments by the original author.

With this commit I have started to understand just how I need to adapt the script
to curve along the Z-axis as opposed to the Y-axis as it originally did. Additionally the
text itself must be rotated along the Y-axis in order to be readable when the text is curved.
There is also something that makes the text only utilise half of the AnimationCurve, which I
have to figure out.

Although there is a lot of concepts to understand, this is a solid step in the right direction.
There are some advanced math concepts to grasp here for someone who has not had any courses within
the game programming disipline. (Dot product, Cross product, and so on).

End of day commit, but with a smile this time.