Added the original form of the script again.
Seems like the curve worked "as expected", producing a full curve in the original.
Problem however is that the difference between the original implmentation and the one
I have been attempting is that the curve is duplicated in the original, meaning the
curve is still cut off half way through. Duplicating the curve made it seem like the
entire curve was being used to draw the text on...