Different approach to text curving
Instead of programatically trying to calculate the angle of the text along
a curve and the rotation of each letter according to the curve, this is an attempt
through creating 3D assets in blender and scaling them up in Unity in order to
scale the text size up and down. It doesnt seem like much quality is lost in scaling up
the text, which means this could potentially be a much more efficient way to implement this.
There's no point going for perfection if it's not noticeable for the end user and this method
also avoids having to buy 3rd party solutions to text curving in Unity.

Main drawback is not being able to programatically change the text content, but this might not
be a problem, depending on how many different scenarios the text needs to be suited for. I assume
there will not be more than 10 different blocks of text, which should not take as much time as
getting an example of curving text programatically has done so far..