Script to make curved text meshes
Created a script for creating curved text meshes to the point where it can now
create a static piece of text. Further work will make it so that the script can
quickly create curved text meshes with different text sizes, circle radiuses and
text box widths.

Added a mesh from the model that displays the results from the script at
0.5 to 2 meters radius(scale) and box size (scale*1.5)