Moved folder for blender scripts and added functionality to the mesh generator
The script is now able to create models based on the JSON file that is used by
TextMeshPro. This means that by running the script all models can be created with
the texts found in the json file with a given radius and textBoxWidth.

Additionally I have refactored a lot of the code into methods, as I have started
to polish the script because I believe it is almost finished.