Generated curved text mesh and made a folder for fonts
Just another generation of the different text meshes, so that
I can more easily view them. Also made a folder for font files
that are imported from the Windows/Fonts folder, so that fonts
stay consistent in Blender and Unity. I have only used Arial so far.
as that is must likely the font I will go for with the experiment