Reworked the entire scene
Remade the entire scene with assets from a low poly pack fetched from the asset store.
The intention is making somewhat of a museum to provide context for the readability tasks
Besides the environment, I have created some glass display cases myself, as I could not find
any free models to use. I have put an Elk and a Parrot inside some display cases to demonstrate
their use and possibly for use in experiments.

Currently I'm trying to work on baking a lightmap into the scene, but there seems to be problems
when there is a lot of objects involved. Unity kept crashing when baking at some point, so I then
realized I have been working on this for way too long without pushing to remote..
As I managed to stop the crashing for now this is my backup :)