Updated python script to group objects generated at each radius
Made it so the python script now groups every textObject with different
widths under the circle (as a parent). Hopefully that will make it easier
to iterate over the collections of objects.

Additionally gave each object generated a name so it is easier to identify
what kind of parameters the text in unity was created with