Add possibility for changing between meshes while in VR
Made it so that all the generated meshes from the python script
can now be viewed in VR. This means hiding/showing only one
mesh at a time and make sure it loops correctly (might be changed
to simply be a max value and min value where it stops, as opposed to
looping).

First I thought I should also increase the radiuses of the circles themselves,
but that might not be the best approach since I can just scale up the
collection of meshes directly in Unity and it seems to work.

Instead I spend those verticies to generate way more examples with
different textBoxWidths.

A lot of time was also spent naming the meshes correctly, as the export
to Unity seems to care about numbering for ordering..