Finished functionality for changing which mesh is visible
Spent a lot of time trying to get the meshes that is children
of the generated object from Blender to loop in terms of visibility.
However, this is not what I actually needed, since it would make the
text jump from the MAX amount of text width the MIN amount of textwidth.
Instead I have now made it so that when you reach either end of the list,
you are no longer allowed to go further in that same direction, you are
however allowed to change direction again, keeping you in the available range.

Also generated text from 0.8 textbox width to 3.0 textboxwidth, as I
believe this is  a good range for now considering how many characters
per line this allowed as MIN and MAX.