Small improvements to file handling
StreamWriter(..., true) is supposed to append to a file, but it
doesn't work as I expected. Instead I'm using File.AppendText().

Couldnt get the participantId to be stored in the FileHandler class,
as it would not increment properly which in turn made it so new files
were never created after the initial file. Hmm.