Continued work on the formatting of experiment data.
This is a continuation of formatting experiment data. To keep
a structure the class ExperimentData was constructed, so that
I could easier keep track of the data when going from the data
being sent to it being written to file. I chose to go for .csv
format as I believe this is a good format when importing the data
later to SPSS.

Initially I thought that I would simply build up a giant ExperimentData
object with all the different data stored in it, but I see that
a lot of the fields would be fairly similar. Instead of doing that
I think it would be much more readable and managable to make a field
that differantiates between different values. E.g. are we reporting
data about the flat screen right now at a comfortable distance
or are we trying to find the minimum readable distance?
Instead of having all the participant position and screen vectors
all thrown into the same line of data in the .csv, I think placing it
on two lines with a "ExperimentId" to differantiate is better.
I will however discuss this with my supervisor.
I have not made it so that each participant has multiple lines connected
to them in this commit, but this is a point I will implement after the
discussion.

I also revisited the background for the flat text, because I thought
maybe the use of a thin cube was the reason for text flickering with
white text and black background. The option I tried out was a canvas
with a panel instead, but the flickering still seems to persist.
This is however probably the better approach anyways, so I might keep
it and delete the thin cube.

End of day commit.