Finished setting up the data structure to write one line per participant.
Finished setting up the structure for the ExperimentData class
and how to handle each individual field in the Filehandler.
Everything is written in .csv format with a single line per
participant.

Next step will be to make it so that each value if reported correctly
when the participants use the controllers in the different experiment stages.