Finished setting up experiment stages and data formatting for now.
Split up the logic for deciding which screen to be showed and
which experiment stage we are in to be two separate Enums.
Made it so the correct data is put into the correct ExperimentData
variable.

The sequences of control on my side can be automated further.
E.g. when I change ExperimentStage to the curved screens, the screen
should automatically switch.