Flat screen now scales with text size and added width changing functionality
Made it so that the scale of the screen behind the text should now scale when
the text size is increased. Made it so that the line width can be adjusted
separately, similar to how the line width changes with the curved screen.

Will have to set up more experiment stages if I want to measure the line width
in addition, perhaps leave it at just ask for a comfortable line width.
Another day another question.