Added calculations for curvedScreen values and added them to the file writer
Made it so that the values for lineHeight, angular size and DMM
is now written to file flr the curved screen as it would be for the flat screen.

Also corrected an error in the calculation for distance to screen, as it
were written so that negative values would not be substracted properly.
For some reason I had previously decided that I needed to substract positive
values only, but that would make it so that it reported the user being closer
to the screen if positioned in negative z-values.