Added fields needed to store lineWidth variables and control setup.
Added necessary fields needed to store values when investigating
the lineWidth people prefer after they have found a comfortable
text size. This includes disabling controls for scaling when
lineWidth is to be adjusted and only have lineWidth adjusting
available in this stage. Need to merge the branch flatscreen-scaling
in order to get all this functionality, but the control scheme is ready.