LineWidth data revisited, bugs quashed and beta run
Updated the model for curved text, as I could still fit in
more different line widths. In addition I made it so text
would be added the longer the line widths, with multiple stages
as opposed to only having one breaking point.

Shouldn't have done the beta run in this branch, but thats what
we got. It seems like the value for moving the curved screen was
changed by merging, which I did not discover prior to testing, making
it so that the participant was closer to the screen than intended.
This does not really matter in this case, since every value is calculated
with the relative distance, but it gives values that are harder to compare
to the flat screen.